+++ .warp/setup/networks/networks.sh
@@ -14,7 +14,7 @@ sleep 1
         B="$(echo $HTTP_HOST_IP | cut -f2 -d . )"
         C="$(echo $HTTP_HOST_IP | cut -f3 -d . )"
         
-        echo "# Network information" >> $ENVIRONMENTVARIABLESFILESAMPLE
+        #echo "# Network information" >> $ENVIRONMENTVARIABLESFILESAMPLE
         echo "NETWORK_SUBNET=$A.$B.$C.0/24" >> $ENVIRONMENTVARIABLESFILESAMPLE
         echo "NETWORK_GATEWAY=$A.$B.$C.1" >> $ENVIRONMENTVARIABLESFILESAMPLE
         echo "NETWORK_NAME=$NETWORK_NAME" >> $ENVIRONMENTVARIABLESFILESAMPLE
@@ -22,7 +22,7 @@ sleep 1
         warp_network_multi
     else
 
-        echo "# Network information" >> $ENVIRONMENTVARIABLESFILESAMPLE
+        #echo "# Network information" >> $ENVIRONMENTVARIABLESFILESAMPLE
         echo "NETWORK_SUBNET=0.0.0.0/24" >> $ENVIRONMENTVARIABLESFILESAMPLE
         echo "NETWORK_GATEWAY=0.0.0.0" >> $ENVIRONMENTVARIABLESFILESAMPLE
         echo "NETWORK_NAME=$NETWORK_NAME" >> $ENVIRONMENTVARIABLESFILESAMPLE